+++ CHANGELOG.md
@@ -11,6 +11,7 @@ IMPROVEMENTS:
   - Change NDID APIs (`/ndid/update_node`, `/ndid/enable_node`, and  `/ndid/disable_node`) HTTP success response status code from 200 to 204.
   - New API: GET `/idp/request_message_padded_hash` for getting `request_message_padded_hash` used for creating mode 2 and 3 request signature (signed with accessor private key without padding) on response.
 - Add accessor in request reference group validation on IdP responses.
+- Support Tendermint 0.32 (Block result spec change).
 
 BUG FIXES: